MODEXPS-26 - Improvement

+++ src/main/java/org/folio/des/scheduling/ExportScheduler.java
@@ -35,9 +35,11 @@ public class ExportScheduler implements SchedulingConfigurer {
     registrar = taskRegistrar;
     taskRegistrar.setScheduler(Executors.newScheduledThreadPool(100));
     taskRegistrar.addTriggerTask(() -> {
-      log.info("configureTasks: is module registered: {}", contextHelper.isModuleRegistered());
-      contextHelper.initScope();
-      jobService.upsert(scheduledJob);
+      log.info("configureTasks: is module registered: {} for tenant: {}", contextHelper.isModuleRegistered(), contextHelper.getTenantId());
+      if (contextHelper.isModuleRegistered()) {
+        contextHelper.initScope();
+        jobService.upsert(scheduledJob);
+      }
     }, trigger);
   }
 
@@ -53,9 +55,11 @@ public class ExportScheduler implements SchedulingConfigurer {
 
   @Scheduled(fixedRateString = "P1D")
   public void deleteOldJobs() {
-    log.info("deleteOldJobs: is module registered: {}", contextHelper.isModuleRegistered());
-    contextHelper.initScope();
-    jobService.deleteOldJobs();
+    log.info("deleteOldJobs: is module registered: {} for tenant: {}", contextHelper.isModuleRegistered() , contextHelper.getTenantId());
+    if (contextHelper.isModuleRegistered()) {
+      contextHelper.initScope();
+      jobService.deleteOldJobs();
+    }
   }
 
   private void reconfigureSchedule() {
